Original change: https://android-review.googlesource.com/c/platform/tools/tradefederation/+/2899374

+++ src/com/android/tradefed/util/image/DeviceImageTracker.java
@@ -135,12 +135,17 @@ public class DeviceImageTracker {
         FileUtil.deleteFile(copyInCacheDeviceImage);
         File copyInCacheBootloader = new File(mCacheDir, serial + "_bootloader");
         FileUtil.deleteFile(copyInCacheBootloader);
-        File copyInCacheBaseband = new File(mCacheDir, serial + "_baseband");
-        FileUtil.deleteFile(copyInCacheBaseband);
+        File copyInCacheBaseband = null;
+        if (baseband != null) { // Baseband is optional on some devices
+            copyInCacheBaseband = new File(mCacheDir, serial + "_baseband");
+            FileUtil.deleteFile(copyInCacheBaseband);
+        }
         try {
             FileUtil.hardlinkFile(deviceImage, copyInCacheDeviceImage);
             FileUtil.hardlinkFile(bootloader, copyInCacheBootloader);
-            FileUtil.hardlinkFile(baseband, copyInCacheBaseband);
+            if (copyInCacheBaseband != null) {
+                FileUtil.hardlinkFile(baseband, copyInCacheBaseband);
+            }
             mImageCache.put(
                     serial,
                     new FileCacheTracker(
